feat: add Stacker contract (#64)
* feat: add Stacker contract

* fix test

* add stacker trait

* add fns to stacker trait

* add extra method and test

* add method to request withdrawal of STX tokens for vaults

* add more stacker trait methods

* Clear TODO

* fix test

* add asserts

* move stacker variables into stacker contract

* fix static call
