+++ clarity/contracts/stx-reserve.clar
@@ -47,6 +47,16 @@
   )
 )
 
+;; transfers (var-get tokens-to-stack) tokens to the stacker contract
+(define-public (request-stx-to-stack)
+  (begin
+    (asserts! (is-eq contract-caller (unwrap-panic (contract-call? .dao get-qualified-name-by-name "stacker"))) (err ERR-NOT-AUTHORIZED))
+    (as-contract
+      (stx-transfer? (var-get tokens-to-stack) (as-contract tx-sender) (unwrap-panic (contract-call? .dao get-qualified-name-by-name "stacker")))
+    )
+  )
+)
+
 ;; calculate the amount of stablecoins to mint, based on posted STX amount
 ;; ustx-amount * stx-price-in-cents == dollar-collateral-posted-in-cents
 ;; (dollar-collateral-posted-in-cents / collateral-to-debt-ratio) == stablecoins to mint
