+++ clarity/contracts/vault-manager-trait.clar
@@ -4,7 +4,6 @@
 (define-trait vault-manager-trait
   (
     (fetch-vault-by-id (uint) (response (tuple (id uint) (owner principal) (collateral uint) (collateral-type (string-ascii 12)) (collateral-token (string-ascii 12)) (stacked-tokens uint) (revoked-stacking bool) (debt uint) (created-at-block-height uint) (updated-at-block-height uint) (stability-fee uint) (stability-fee-last-accrued uint) (is-liquidated bool) (auction-ended bool) (leftover-collateral uint)) bool))
-    (get-stacking-unlock-burn-height () (response uint bool))
     (get-stx-redeemable () (response uint bool))
     (get-last-vault-id () (response uint uint))
